@@ -1,4 +1,5 @@
 import React, { Component } from "react";
+import PropTypes from 'prop-types';
 import { ReactiveBase, RatingsFilter, ResultCard, ResultList, ViewSwitcher } from "../app";
 import ResponsiveStory from "./ResponsiveStory";
 
@@ -139,8 +140,8 @@ ViewSwitcherDefault.defaultProps = {
 };
 
 ViewSwitcherDefault.propTypes = {
-	mapping: React.PropTypes.shape({
-		rating: React.PropTypes.string,
-		name: React.PropTypes.string
+	mapping: PropTypes.shape({
+		rating: PropTypes.string,
+		name: PropTypes.string
 	})
 };
